pyproject.toml
[project]
name = "agentdemo"
version = "1.0.0"
description = "Weather-Based Clothing Advisor - POC demonstrating dual deployment (Container Apps + Foundry)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Azure Agent Framework (https://github.com/microsoft/agent-framework)
    "agent-framework-core",
    "agent-framework-azure-ai",

    # Azure AI Foundry (for Foundry deployment)
    "azure-ai-projects>=1.0.0",
    "azure-identity>=1.15.0",

    # Async utilities
    "nest-asyncio>=1.5.0",

    # FastAPI (for Container Apps deployment)
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",

    # HTTP client
    "requests>=2.31.0",

    # YAML parsing (for Foundry workflow deployment)
    "pyyaml>=6.0.1",

    # Pydantic (for Agent Framework tool definitions)
    "pydantic>=2.0.0",

    # Telemetry
    "azure-monitor-opentelemetry>=1.2.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
]
